Fix for delay in entering recovery mode after FDR.
There is 2~3 mins to enter into recovery mode when
execute FDR. There is delay in starting FDR Service.
The process is not running at a high priority.

Increase the number of background services and make
sure that FDR service process will run at high
priority in order to reduce the delay to start FDR.

CRs-Fixed: 802810

Change-Id: I96a55d8d99f942b3f597bcc82f55ef38a98dd6c2