msm8226: Add a property to restrict number of cameras to 2
In L AOSP MAX camera defined as 4, compared to 2 in KK
Due to that and multimode architecture, camera switching on 8x26 target
will fail on KK-based camera as there are non-zero specific profile
exists in dtsi file. These profiles are needed for JB-based camera.
With this setproperty, HAL will restrict number of cameras to two,
if there are more than 2 valid profiles.
Change-Id: I9c76a27d348d3f02371e1767c9230ae50c88f99e